refactored btc
