+++ Bitcoin/cron/cronservice.js
@@ -6,6 +6,8 @@ var TxServiceInofModel = require("../model/txServiceInfo");
 var AddressModel = require("../model/address");
 var AddrServiceInofModel = require("../model/addrServiceInfo");
 
+var UtilsModule = require("../modules/utils");
+
 var fs = require("fs");
 var Log = require("log"),
   log = new Log(
@@ -15,25 +17,10 @@ var Log = require("log"),
 // log = new Log('debug');
 
 function filelog(...params) {
-  //
   log.info(...params);
-  // console.log(...params);
 }
 
-var promisify = function promisify(fn, args) {
-  return new Promise((resolve, reject) => {
-    try {
-      client.call(fn, args, function(err, result) {
-        if (err) {
-          reject(err);
-        }
-        resolve(result);
-      });
-    } catch (error) {
-      reject(error);
-    }
-  });
-};
+var promisify = UtilsModule.promisify;
 
 ////////////////////////////////////////////////////////////////////////////
 
@@ -175,39 +162,39 @@ async function CheckUpdatedAddresses() {
       let { txid } = arrTxs[i];
       try {
         var txInfo;
-        // if (txid == '4a5e1e4baab89f3a32518a88c31bc87f618f76673e2cc77ab2127b7afdeda33b') {
-        //   txInfo = {
-        //     "txid": "4a5e1e4baab89f3a32518a88c31bc87f618f76673e2cc77ab2127b7afdeda33b",
-        //     "version": 1,
-        //     "locktime": 0,
-        //     "vin": [
-        //       {
-        //         "coinbase": "04ffff001d0104455468652054696d65732030332f4a616e2f32303039204368616e63656c6c6f72206f6e206272696e6b206f66207365636f6e64206261696c6f757420666f722062616e6b73",
-        //         "sequence": 4294967295
-        //       }
-        //     ],
-        //     "vout": [
-        //       {
-        //         "value": 50.00000000,
-        //         "n": 0,
-        //         "scriptPubKey": {
-        //           "asm": "04678afdb0fe5548271967f1a67130b7105cd6a828e03909a67962e0ea1f61deb649f6bc3f4cef38c4f35504e51ec112de5c384df7ba0b8d578a4c702b6bf11d5f OP_CHECKSIG",
-        //           "hex": "4104678afdb0fe5548271967f1a67130b7105cd6a828e03909a67962e0ea1f61deb649f6bc3f4cef38c4f35504e51ec112de5c384df7ba0b8d578a4c702b6bf11d5fac",
-        //           "reqSigs": 1,
-        //           "type": "pubkey",
-        //           "addresses": [
-        //             "1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa"
-        //           ]
-        //         }
-        //       }
-        //     ]
-        //   }
-        // } else {
-        //   txInfo = await promisify("getrawtransaction", [txid, 1]);
-        // }
         try {
           txInfo = await promisify("getrawtransaction", [txid, 1]);
         } catch (error) {
+          // if (txid == '4a5e1e4baab89f3a32518a88c31bc87f618f76673e2cc77ab2127b7afdeda33b') {
+          //   txInfo = {
+          //     "txid": "4a5e1e4baab89f3a32518a88c31bc87f618f76673e2cc77ab2127b7afdeda33b",
+          //     "version": 1,
+          //     "locktime": 0,
+          //     "vin": [
+          //       {
+          //         "coinbase": "04ffff001d0104455468652054696d65732030332f4a616e2f32303039204368616e63656c6c6f72206f6e206272696e6b206f66207365636f6e64206261696c6f757420666f722062616e6b73",
+          //         "sequence": 4294967295
+          //       }
+          //     ],
+          //     "vout": [
+          //       {
+          //         "value": 50.00000000,
+          //         "n": 0,
+          //         "scriptPubKey": {
+          //           "asm": "04678afdb0fe5548271967f1a67130b7105cd6a828e03909a67962e0ea1f61deb649f6bc3f4cef38c4f35504e51ec112de5c384df7ba0b8d578a4c702b6bf11d5f OP_CHECKSIG",
+          //           "hex": "4104678afdb0fe5548271967f1a67130b7105cd6a828e03909a67962e0ea1f61deb649f6bc3f4cef38c4f35504e51ec112de5c384df7ba0b8d578a4c702b6bf11d5fac",
+          //           "reqSigs": 1,
+          //           "type": "pubkey",
+          //           "addresses": [
+          //             "1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa"
+          //           ]
+          //         }
+          //       }
+          //     ]
+          //   }
+          // } else {
+          //   txInfo = await promisify("getrawtransaction", [txid, 1]);
+          // }
           filelog(
             `error in getrawtransaction: i=${i}, txid=${txid}, lasttxid:${lastTxid}, lasttxoffset:${lastTxOffset}`,
             error
@@ -225,8 +212,10 @@ async function CheckUpdatedAddresses() {
               if (!inTxid || inTxid == "" || inVout < 0) continue;
               var inTxInfo = await promisify("gettxout", [inTxid, inVout]);
               if (!inTxInfo) {
+                filelog(
+                  `inTxInfo not found. inTxid=${inTxid}, vout: ${inVout}`
+                );
                 continue;
-                // throw `inTxInfo not found. inTxid=${inTxid}`;
               }
 
               var addresses = inTxInfo.scriptPubKey.addresses;
@@ -238,19 +227,26 @@ async function CheckUpdatedAddresses() {
                 if (!addressRow) {
                   addressRow = new AddressModel({
                     address: addresses[k],
-                    txs: []
+                    txs: [],
+                    txsIn: [],
+                    txsOut: []
                   });
                 }
                 if (addressRow.txs.indexOf(txid) == -1) {
                   addressRow.txs.push(txid);
-                  try {
-                    await addressRow.save();
-                  } catch (e) {
-                    filelog(
-                      `vin addressRow.save.txs error: i=${i}, j=${j}, k=${k}, addressRow=${addressRow}`
-                    ); // Should dump errors here
-                    throw e;
-                  }
+                }
+                if (
+                  addressRow.txsIn.indexOf({ txid: inTxid, vout: inVout }) == -1
+                ) {
+                  addressRow.txsIn.push({ txid: inTxid, vout: inVout });
+                }
+                try {
+                  await addressRow.save();
+                } catch (e) {
+                  filelog(
+                    `vin addressRow.save.txs vin error: i=${i}, j=${j}, k=${k}, addressRow=${addressRow}`
+                  ); // Should dump errors here
+                  throw e;
                 }
               }
             }
@@ -272,16 +268,19 @@ async function CheckUpdatedAddresses() {
                   }
                   if (addressRow.txs.indexOf(txid) == -1) {
                     addressRow.txs.push(txid);
-                    try {
-                      await addressRow.save();
-                    } catch (e) {
-                      filelog(
-                        `addressRow.save.txs error: i=${i}, j=${j}, k=${k} addressRow=${addressRow}`
-                      ); // Should dump errors here
-                      throw e;
-                    }
                   }
-                }
+                  if (addressRow.txsOut.indexOf({txid, vout: j}) == -1) {
+                    addressRow.txsOut.push({txid, vout: j});
+                  }
+                  try {
+                    await addressRow.save();
+                  } catch (e) {
+                    filelog(
+                      `addressRow.save.txs vout error: i=${i}, j=${j}, k=${k} addressRow=${addressRow}`
+                    ); // Should dump errors here
+                    throw e;
+                  }
+              }
               }
             }
           }
@@ -299,17 +298,13 @@ async function CheckUpdatedAddresses() {
 }
 
 async function transactionService() {
-  filelog("Start transaction service ...");
   await CheckUpdatedTransactions();
   setTimeout(transactionService, config.TX_CRON_TIME);
-  filelog("Done transaction service .");
 }
 
 async function addressService() {
-  filelog("Start address service ...");
   await CheckUpdatedAddresses();
   setTimeout(addressService, config.ADDR_CRON_TIME);
-  filelog("Done address service .");
 }
 
 exports.start_cronService = async function() {
